reusableRDBMS DAO. Add getAll operation, unify DAO interfaces, move notifications to composite DAO

+++ storage/src/test/java/pl/edu/icm/unity/store/IdentityTypeTest.java
@@ -34,7 +34,8 @@ public class IdentityTypeTest
 	public void shouldReturnCreatedIdentityType() throws EngineException
 	{
 		tx.runInTransaction(() -> {
-			idTypeDAO.delete(MockIdentityTypeDef.NAME);
+			if (idTypeDAO.exists(MockIdentityTypeDef.NAME))
+				idTypeDAO.delete(MockIdentityTypeDef.NAME);
 		});
 		
 		IdentityType idType = new IdentityType(new MockIdentityTypeDef());
